# perf(fslib): execute RegExp once when converting to/from `PortablePath`s