… packages/yarnpkg-core/sources/worker-zip/index.js
… packages/yarnpkg-pnp/sources/hook.js