**What's the problem this PR addresses?**

Converting to/from `PortablePath`s on Windows executes the same RegExp twice instead of reusing the result

**How did you fix it?**

Reuse the result

**Result**
Tested on the Storybook repo using PnP

```
YARN_IGNORE_PATH=1 hyperfine \
                  "node ./master.cjs --mode skip-build" \
                  "node ./after.cjs --mode skip-build"
Benchmark #1: node ./master.cjs --mode skip-build
  Time (mean ± σ):      3.495 s ±  0.048 s    [User: 4.4 ms, System: 4.2 ms]
  Range (min … max):    3.454 s …  3.585 s    10 runs

Benchmark #2: node ./after.cjs --mode skip-build
  Time (mean ± σ):      3.461 s ±  0.010 s    [User: 0.0 ms, System: 4.2 ms]
  Range (min … max):    3.450 s …  3.475 s    10 runs

Summary
  'node ./after.cjs --mode skip-build' ran
    1.01 ± 0.01 times faster than 'node ./master.cjs --mode skip-build'
```

**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I will check that all automated PR checks pass before the PR gets reviewed.